backend/requirements.txt
flask
psycopg2-binary>=2.9
python-dotenv
flask-cors